polrev/pyproject.toml
[tool.poetry]
name = "polrev"
version = "0.1.0"
description = ""
authors = ["Kurtis Fields <kurtisfields@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^3.2.12"
wagtail = "^2.15.2"
django-colorful = "^1.3"
django-el-pagination = "^3.3.0"
django-social-share = "^2.2.1"
puput = "^1.1.3"
django-allauth = "^0.47.0"
django-crispy-forms = "^1.13.0"
crispy-bootstrap5 = "^0.6"
django-avatar = "^5.0.0"
django-comments-xtd = "^2.9.4"
wagtail-markdown = "^0.7.0"
lxml = "^4.6.4"
psycopg2-binary = "^2.9.2"
python-dotenv = "^0.19.2"
django-storages = "^1.12.3"
boto3 = "^1.20.21"
django-cors-headers = "^3.10.1"
django-dbbackup = "^3.3.0"
#"ls.joyous" = "^1.4.0"
wagtail-django-recaptcha = "^1.0"
wagtail-birdsong = "^0.2.0"
django-bootstrap5 = "^21.3"
django-tz-detect = "^0.3.0"
tzdata = "^2021.5"
uvicorn = "^0.17.4"
gunicorn = "^20.1.0"
wagtail-autocomplete = "^0.7.0"
django-timezone-field = "^5.0"
num2words = "^0.5.10"
holidays = "^0.13"
icalendar = "^4.0.9"
feedgen = "^0.9.0"
django-model-utils = "^4.2.0"
us = "^2.0.2"
wagtail-generic-chooser = "^0.3.1"
chardet = "^4.0.0"

[tool.poetry.dev-dependencies]
pyppeteer = "^0.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
